requirements.txt
aiogram==2.11.2
aiohttp==3.7.2
APScheduler==3.6.3
async-timeout==3.0.1
atomicwrites==1.4.0
attrs==20.2.0
Babel==2.9.0
boto3==1.16.13
botocore==1.19.13
certifi==2020.6.20
cffi==1.14.3
chardet==3.0.4
colorama==0.4.4
cryptography==3.2.1
decorator==4.4.2
idna==2.10
iniconfig==1.1.1
jmespath==0.10.0
multidict==5.0.0
numpy==1.19.3
packaging==20.4
pluggy==0.13.1
py==1.9.0
pycparser==2.20
pycryptodome==3.9.9
pyparsing==2.4.7
pyTelegramBotAPI==3.7.3
pytest==6.1.2
python-dateutil==2.8.1
python-telegram-bot==13.0
pytz==2020.1
requests==2.7.0
s3transfer==0.3.3
secret==0.8
six==1.15.0
tabulate==0.8.7
telebot==0.0.3
telepot==12.7
toml==0.10.2
tornado==6.1
typing-extensions==3.7.4.3
tzlocal==2.1
urllib3==1.25.11
yarl==1.6.2
